% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_dev.R
\name{combine_dev}
\alias{combine_dev}
\title{Combining dev-versions into the next version}
\usage{
combine_dev(text = "")
}
\arguments{
\item{text}{a string vector with the content of the get_text method}
}
\value{
a string vector with the reduced NEWS.md text
}
\description{
Combining dev-versions into the next version
}
