% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsmd.R
\docType{class}
\name{news}
\alias{news}
\title{Manipulate the NEWS.md file}
\format{An R6 class.}
\usage{
news
}
\description{
Manipulate the NEWS.md file.
}
\examples{
## Create a template
my_news <- news$new()
my_news$add_subtitle("improved things 1")
my_news$add_bullet("point 1")
my_news$add_bullet("point 2")

}
\seealso{
\link{newsmd}
}
\keyword{datasets}
