% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsmd.R
\name{newsmd}
\alias{newsmd}
\title{Creating a NEWS.md file}
\usage{
newsmd(
  file = NULL,
  text = c(paste0("## version ", version), "", "---", "", "### NEWS.md setup", "",
    "- added NEWS.md creation with [newsmd](https://github.com/Dschaykib/newsmd)", ""),
  version = "0.0.0.9000"
)
}
\arguments{
\item{file}{name of the \code{NEWS} file to load. If not given a new
file \code{NEWS.md} is created.}

\item{text}{a character scalar containing the initial text.}

\item{version}{a character with the version in the format
\code{major.minor.patch.dev}.}
}
\description{
This is a convenience wrapper for \code{news$new()}. If you already have a
 \code{NEWS} file you can add it's path to \code{newsmd()}.
}
\details{
This functions creates a \code{news} object, which can be saved as the
 NEWS.md file with the internal method \code{write()}. One can add versions,
 subtitles and bullet points to the \code{news}.

 If an existing NEWS.md file is given, the version is extracted with
 \code{get_version} and the dev part is bumped up.
}
\examples{
newsmd()
}
